Only pass mountpoint's ID to service's delete function

+++ byceps/services/snippet/mountpoint_service.py
@@ -29,8 +29,13 @@ def create_mountpoint(
     return mountpoint
 
 
-def delete_mountpoint(mountpoint: DbMountpoint) -> None:
+def delete_mountpoint(mountpoint_id: MountpointID) -> None:
     """Delete the mountpoint."""
+    mountpoint = find_mountpoint(mountpoint_id)
+
+    if mountpoint is None:
+        raise ValueError(f"Unknown mountpoint ID '{mountpoint_id}'.")
+
     db.session.delete(mountpoint)
     db.session.commit()
